Refactor trace store index helpers
